refactor: move views
