--- lib/views/doc.go
@@ -1,2 +1,0 @@
-// Package views はエンティティを組み込んだ、具体的なUIの部品を提供する。
-package views
