+++ lib/widgets/views/spec.go
